libutils: Mmio + refactoring
Splitting up the io module in libutils in three modules:

* mod.rs: Contains the Io trait, and ReadOnly/WriteOnly wrappers,
* pio.rs: Contains the Port IO implementation for x86,
* mmio.rs: New module, contains interface to a Memory Mapped IO.
